Should provide a slightly cleaner looking logo
I didn't like how the link for the logo was stretched wide and super short, it made it so if you moused over the box in the center of the site title's border it wouldn't always pick up the mouse. This should make it so the link takes up the whole box, it does also make the link itself shorter but only to the size of the image inside.